requirements.txt
fastapi==0.88.0
uvicorn==0.20.0
sqlalchemy==1.4.44
pydantic==1.10.2
marshmallow==3.19.0 
marshmallow_sqlalchemy==0.28.1
psycopg2==2.8.4
psycopg2-binary==2.9.5
alembic==1.8.1
pytest==7.2.0
httpx==0.23.1
PyJWT==2.6.0
cryptograpy==38.0.4